@@ -1,34 +1,24 @@
 cmake_minimum_required(VERSION 3.5)
 project(mrs_uav_state_estimators)
 
-set(CATKIN_DEPENDENCIES
-  cmake_modules
-  dynamic_reconfigure
-  geometry_msgs
-  message_generation
-  message_runtime
-  mrs_lib
-  mrs_msgs
-  mrs_uav_managers
-  nav_msgs
-  nodelet
-  roscpp
-  roslint
-  sensor_msgs
-  std_msgs
-  tf2_geometry_msgs
-  mrs_uav_testing
-  )
-
-find_package(catkin REQUIRED COMPONENTS
-  ${CATKIN_DEPENDENCIES}
-  )
-
-set(CMAKE_CXX_STANDARD 17)
+# set the correct standards
+set(CMAKE_C_STANDARD 99)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
-add_definitions(-Wall)
-add_definitions(-Wextra)
+# set the compile options to show code warnings
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra)
+endif()
+
+if(MRS_ENABLE_TESTING)
+  message(WARNING "Setting optimizaton levels for tests")
+  add_definitions("-O0")
+  add_definitions("-g")
+else()
+  add_definitions("-O2")
+endif()
 
 if(COVERAGE)
   message(WARNING "building with --coverage, the performance might be limited")
