@@ -4,6 +4,8 @@
 
 #include <mrs_uav_state_estimators/estimators/heading/hdg_generic.h>
 
+#include <ament_index_cpp/get_package_share_directory.hpp>
+
 //}
 
 namespace mrs_uav_state_estimators
@@ -11,7 +13,10 @@ namespace mrs_uav_state_estimators
 {
 
 /* initialize() //{*/
-void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) {
+void HdgGeneric::initialize(const rclcpp::Node::SharedPtr &node, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) {
+
+  node_  = node;
+  clock_ = node->get_clock();
 
   ch_ = ch;
   ph_ = ph;
@@ -29,18 +34,18 @@ void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
   H_ <<
     1, 0;
 
-  innovation_ << 
+  innovation_ <<
     0;
-
-
   // clang-format on
 
   // | --------------- initialize parameter loader -------------- |
 
   if (is_core_plugin_) {
 
-    ph->param_loader->addYamlFile(ros::package::getPath(package_name_) + "/config/private/" + parent_state_est_name_ + "/" + getName() + ".yaml");
-    ph->param_loader->addYamlFile(ros::package::getPath(package_name_) + "/config/public/" + parent_state_est_name_ + "/" + getName() + ".yaml");
+    ph->param_loader->addYamlFile(ament_index_cpp::get_package_share_directory(package_name_) + "/config/private/" + parent_state_est_name_ + "/" + getName() +
+                                  ".yaml");
+    ph->param_loader->addYamlFile(ament_index_cpp::get_package_share_directory(package_name_) + "/config/public/" + parent_state_est_name_ + "/" + getName() +
+                                  ".yaml");
   }
 
   ph->param_loader->setPrefix(ch_->package_name + "/" + Support::toSnakeCase(ch_->nodelet_name) + "/" + getNamespacedName() + "/");
@@ -58,7 +63,7 @@ void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
   for (auto corr_name : correction_names_) {
     corrections_.push_back(std::make_shared<Correction<hdg_generic::n_measurements>>(
-        nh, getNamespacedName(), corr_name, ns_frame_id_, EstimatorType_t::HEADING, ch_, ph_, [this](int a, int b) { return this->getState(a, b); },
+        node, getNamespacedName(), corr_name, ns_frame_id_, EstimatorType_t::HEADING, ch_, ph_, [this](int a, int b) { return this->getState(a, b); },
         [this](const Correction<hdg_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t state) {
           return this->doCorrection(meas, R, state);
         }));
@@ -76,16 +81,49 @@ void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
   // | ------- check if all parameters loaded successfully ------ |
   if (!ph->param_loader->loadedSuccessfully()) {
-    ROS_ERROR("[%s]: Could not load all non-optional parameters. Shutting down.", getPrintName().c_str());
-    ros::shutdown();
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: Could not load all non-optional parameters. Shutting down.", getPrintName().c_str());
+    rclcpp::shutdown();
   }
 
   // | ------------- initialize dynamic reconfigure ------------- |
-  drmgr_ =
-      std::make_unique<drmgr_t>(ros::NodeHandle("~/" + getNamespacedName()), true, getPrintName(), boost::bind(&HdgGeneric::callbackReconfigure, this, _1, _2));
-  drmgr_->config.pos = Q_(POSITION, POSITION);
-  drmgr_->config.vel = Q_(VELOCITY, VELOCITY);
-  drmgr_->update_config(drmgr_->config);
+
+  // original ROS1 dynamic reconfigure server
+  /* drmgr_ = */
+  /*     std::make_unique<drmgr_t>(ros::NodeHandle("~/" + getNamespacedName()), true, getPrintName(), std::bind(&HdgGeneric::callbackReconfigure, this, _1,
+   * _2)); */
+  /* drmgr_->config.pos = Q_(POSITION, POSITION); */
+  /* drmgr_->config.vel = Q_(VELOCITY, VELOCITY); */
+  /* drmgr_->update_config(drmgr_->config); */
+
+  {
+    auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+
+    rcl_interfaces::msg::FloatingPointRange range;
+
+    range.from_value = 0.0;
+    range.to_value   = 100000.0;
+
+    param_desc.floating_point_range = {range};
+
+    param_desc.type = rcl_interfaces::msg::ParameterType::PARAMETER_DOUBLE;
+
+    node_->declare_parameter("pos", 0.0, param_desc);
+  }
+
+  {
+    auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+
+    rcl_interfaces::msg::FloatingPointRange range;
+
+    range.from_value = 0.0;
+    range.to_value   = 100000.0;
+
+    param_desc.floating_point_range = {range};
+
+    param_desc.type = rcl_interfaces::msg::ParameterType::PARAMETER_DOUBLE;
+
+    node_->declare_parameter("vel", 0.0, param_desc);
+  }
 
   // | --------------- Kalman filter intialization -------------- |
   const x_t        x0 = x_t::Zero();
@@ -98,47 +136,60 @@ void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
     generateRepredictorModels(input_coeff_);
 
-    const u_t       u0 = u_t::Zero();
-    const ros::Time t0 = ros::Time::now();
-    lkf_rep_           = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
+    const u_t          u0 = u_t::Zero();
+    const rclcpp::Time t0 = clock_->now();
+    lkf_rep_              = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
 
     setDt(1.0 / ch_->desired_uav_state_rate);
   }
 
-  // | ------------------ timers initialization ----------------- |
-  timer_update_ = nh.createTimer(ros::Rate(ch_->desired_uav_state_rate), &HdgGeneric::timerUpdate, this);  // not running after init
-  /* timer_check_health_ = nh.createTimer(ros::Rate(ch_->desired_uav_state_rate), &HdgGeneric::timerCheckHealth, this); */
+  // | ------------------------- timers ------------------------- |
+
+  mrs_lib::TimerHandlerOptions opts;
+
+  opts.node      = node_;
+  opts.autostart = true;
+
+  {
+    std::function<void()> callback_fcn = std::bind(&HdgGeneric::timerUpdate, this);
+
+    timer_update_ = std::make_shared<mrs_lib::ROSTimer>(opts, rclcpp::Rate(ch_->desired_uav_state_rate, clock_), callback_fcn);
+
+    timer_update_last_time_ = rclcpp::Time(0, 0, clock_->get_clock_type());
+  }
 
   // | --------------- subscribers initialization --------------- |
+
   // subscriber to odometry
-  mrs_lib::SubscribeHandlerOptions shopts;
-  shopts.nh                 = nh;
+  mrs_lib::SubscriberHandlerOptions shopts;
+
+  shopts.node               = node;
   shopts.node_name          = getPrintName();
   shopts.no_message_timeout = mrs_lib::no_timeout;
   shopts.threadsafe         = true;
   shopts.autostart          = true;
-  shopts.queue_size         = 10;
-  shopts.transport_hints    = ros::TransportHints().tcpNoDelay();
 
-  sh_control_input_ = mrs_lib::SubscribeHandler<mrs_msgs::EstimatorInput>(shopts, "control_input_in");
+  sh_control_input_ = mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorInput>(shopts, "~/control_input_in");
 
   // | ---------------- publishers initialization --------------- |
   if (ch_->debug_topics.input) {
-    ph_input_ = mrs_lib::PublisherHandler<mrs_msgs::Float64ArrayStamped>(nh, getNamespacedName() + "/input", 10);
+    ph_input_ = mrs_lib::PublisherHandler<mrs_msgs::msg::Float64ArrayStamped>(node, "~/" + getNamespacedName() + "/input");
   }
+
   if (ch_->debug_topics.output) {
-    ph_output_ = mrs_lib::PublisherHandler<mrs_msgs::EstimatorOutput>(nh, getNamespacedName() + "/output", 10);
+    ph_output_ = mrs_lib::PublisherHandler<mrs_msgs::msg::EstimatorOutput>(node, "~/" + getNamespacedName() + "/output");
   }
+
   if (ch_->debug_topics.diag) {
-    ph_diagnostics_ = mrs_lib::PublisherHandler<mrs_msgs::EstimatorDiagnostics>(nh, getNamespacedName() + "/diagnostics", 10);
+    ph_diagnostics_ = mrs_lib::PublisherHandler<mrs_msgs::msg::EstimatorDiagnostics>(node, "~/" + getNamespacedName() + "/diagnostics");
   }
 
   // | ------------------ finish initialization ----------------- |
 
   if (changeState(INITIALIZED_STATE)) {
-    ROS_INFO("[%s]: Estimator initialized, version %s", getPrintName().c_str(), VERSION);
+    RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator initialized, version %s", getPrintName().c_str(), VERSION);
   } else {
-    ROS_INFO("[%s]: Estimator could not be initialized", getPrintName().c_str());
+    RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator could not be initialized", getPrintName().c_str());
   }
 }
 /*//}*/
@@ -147,12 +198,11 @@ void HdgGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 bool HdgGeneric::start(void) {
 
   if (isInState(READY_STATE)) {
-    /* timer_update_.start(); */
     changeState(STARTED_STATE);
     return true;
 
   } else {
-    ROS_WARN_THROTTLE(1.0, "[%s]: Estimator must be in READY_STATE to start it", getPrintName().c_str());
+    RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator must be in READY_STATE to start it", getPrintName().c_str());
     return false;
   }
 }
@@ -175,7 +225,7 @@ bool HdgGeneric::pause(void) {
 bool HdgGeneric::reset(void) {
 
   if (!isInitialized()) {
-    ROS_ERROR("[%s]: Cannot reset uninitialized estimator", getPrintName().c_str());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: Cannot reset uninitialized estimator", getPrintName().c_str());
     return false;
   }
 
@@ -196,20 +246,20 @@ bool HdgGeneric::reset(void) {
   lkf_ = std::make_shared<lkf_t>(A_, B_, H_);
   if (is_repredictor_enabled_) {
 
-    const u_t       u0 = u_t::Zero();
-    const ros::Time t0 = ros::Time(0);
-    lkf_rep_           = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
+    const u_t          u0 = u_t::Zero();
+    const rclcpp::Time t0 = rclcpp::Time(0, 0, clock_->get_clock_type());
+    lkf_rep_              = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
   }
 
   changeState(INITIALIZED_STATE);
-  ROS_INFO("[%s]: Estimator reset", getPrintName().c_str());
+  RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator reset", getPrintName().c_str());
 
   return true;
 }
 /*//}*/
 
 /* timerUpdate() //{*/
-void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
+void HdgGeneric::timerUpdate() {
 
   if (!isInitialized()) {
     return;
@@ -218,12 +268,12 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
   switch (getCurrentSmState()) {
 
     case UNINITIALIZED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
       break;
     }
 
     case READY_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
       break;
     }
 
@@ -234,22 +284,23 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
         if (res) {
           auto measurement_stamped = res.value();
           setState(measurement_stamped.value(AXIS_X), correction->getStateId(), AXIS_X);
-          ROS_INFO_THROTTLE(1.0, "[%s]: Setting initial state to: %.2f", getPrintName().c_str(), measurement_stamped.value(AXIS_X));
+          RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Setting initial state to: %.2f", getPrintName().c_str(),
+                               measurement_stamped.value(AXIS_X));
         } else {
-          ROS_INFO_THROTTLE(1.0, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(),
-                            correction->getNamespacedName().c_str());
+          RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(),
+                               correction->getNamespacedName().c_str());
           return;
         }
       }
-      ROS_INFO_THROTTLE(1.0, "[%s]: Ready to start", getPrintName().c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Ready to start", getPrintName().c_str());
       changeState(READY_STATE);
       break;
     }
 
     case STARTED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
       if (isConverged()) {
-        ROS_INFO_THROTTLE(1.0, "[%s]: LKF converged", getPrintName().c_str());
+        RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: LKF converged", getPrintName().c_str());
         changeState(RUNNING_STATE);
       }
       break;
@@ -258,7 +309,8 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
     case RUNNING_STATE: {
       for (auto correction : corrections_) {
         if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
+          RCLCPP_ERROR_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Correction %s is not healthy!", getPrintName().c_str(),
+                                correction->getNamespacedName().c_str());
           changeState(ERROR_STATE);
         }
       }
@@ -266,16 +318,17 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
     }
 
     case STOPPED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is stopped", getPrintName().c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator is stopped", getPrintName().c_str());
       break;
     }
 
     case ERROR_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
       bool all_corrections_healthy = true;
       for (auto correction : corrections_) {
         if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
+          RCLCPP_ERROR_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Correction %s is not healthy!", getPrintName().c_str(),
+                                correction->getNamespacedName().c_str());
           all_corrections_healthy = false;
         }
       }
@@ -292,9 +345,9 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
   }
 
   // check age of input
-  if (is_input_ready_ && (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec() > 0.1) {
-    ROS_WARN_THROTTLE(1.0, "[%s]: input too old (%.4f), using zero input instead", getPrintName().c_str(),
-                      (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec());
+  if (is_input_ready_ && (clock_->now() - sh_control_input_.lastMsgTime()).seconds() > 0.1) {
+    RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: input too old (%.4f), using zero input instead", getPrintName().c_str(),
+                         (clock_->now() - sh_control_input_.lastMsgTime()).seconds());
     is_input_ready_ = false;
   }
 
@@ -307,7 +360,10 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
     return;
   }
 
-  double dt = (event.current_real - event.last_real).toSec();
+  double dt = (clock_->now() - timer_update_last_time_).seconds();
+
+  timer_update_last_time_ = clock_->now();
+
   if (dt <= 0.0 || dt > 1.0) {
     return;
   }
@@ -326,17 +382,17 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
   /* } */
 
   // prediction step
-  u_t       u;
-  ros::Time input_stamp;
+  u_t          u;
+  rclcpp::Time input_stamp;
   if (is_input_ready_) {
     input_stamp = sh_control_input_.getMsg()->header.stamp;
-    if (input_coeff_ != default_input_coeff_){
+    if (input_coeff_ != default_input_coeff_) {
       setInputCoeff(default_input_coeff_);
     }
     u(0) = sh_control_input_.getMsg()->control_hdg_rate;
   } else {
-    input_stamp = ros::Time::now();
-    if (input_coeff_ != 0){
+    input_stamp = clock_->now();
+    if (input_coeff_ != 0) {
       setInputCoeff(0);
     }
     u = u_t::Zero();
@@ -349,14 +405,14 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
     std::scoped_lock lock(mutex_lkf_);
     if (is_repredictor_enabled_) {
       lkf_rep_->addInputChangeWithNoise(u, Q, input_stamp, models_[0]);
-      sc = lkf_rep_->predictTo(ros::Time::now());
+      sc = lkf_rep_->predictTo(clock_->now());
     } else {
       sc = lkf_->predict(sc, u, Q, dt_);
     }
   }
   catch (const std::exception &e) {
     // In case of error, alert the user
-    ROS_ERROR("[%s]: LKF prediction failed: %s", getPrintName().c_str(), e.what());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: LKF prediction failed: %s", getPrintName().c_str(), e.what());
   }
 
   mrs_lib::set_mutexed(mutex_sc_, sc, sc_);
@@ -372,98 +428,6 @@ void HdgGeneric::timerUpdate(const ros::TimerEvent &event) {
 }
 /*//}*/
 
-/*//{ timerCheckHealth() */
-void HdgGeneric::timerCheckHealth([[maybe_unused]] const ros::TimerEvent &event) {
-
-  if (!isInitialized()) {
-    return;
-  }
-
-  switch (getCurrentSmState()) {
-
-    case UNINITIALIZED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      break;
-    }
-
-    case READY_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      break;
-    }
-
-    case INITIALIZED_STATE: {
-      // initialize the estimator with current corrections
-      for (auto correction : corrections_) {
-        auto res = correction->getProcessedCorrection();
-        if (res) {
-          auto measurement_stamped = res.value();
-          setState(measurement_stamped.value(AXIS_X), correction->getStateId(), AXIS_X);
-          ROS_INFO_THROTTLE(1.0, "[%s]: Setting initial state to: %.2f", getPrintName().c_str(), measurement_stamped.value(AXIS_X));
-        } else {
-          ROS_INFO_THROTTLE(1.0, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(),
-                            correction->getNamespacedName().c_str());
-          return;
-        }
-      }
-      ROS_INFO_THROTTLE(1.0, "[%s]: Ready to start", getPrintName().c_str());
-      changeState(READY_STATE);
-      break;
-    }
-
-    case STARTED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      if (isConverged()) {
-        ROS_INFO_THROTTLE(1.0, "[%s]: LKF converged", getPrintName().c_str());
-        changeState(RUNNING_STATE);
-      }
-      break;
-    }
-
-    case RUNNING_STATE: {
-      for (auto correction : corrections_) {
-        if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
-          changeState(ERROR_STATE);
-        }
-      }
-      break;
-    }
-
-    case STOPPED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is stopped", getPrintName().c_str());
-      break;
-    }
-
-    case ERROR_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
-      bool all_corrections_healthy = true;
-      for (auto correction : corrections_) {
-        if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
-          all_corrections_healthy = false;
-        }
-      }
-      // initialize the estimator again if corrections become healthy
-      if (all_corrections_healthy) {
-        changeState(INITIALIZED_STATE);
-      }
-      break;
-    }
-  }
-
-  if (sh_control_input_.newMsg()) {
-    is_input_ready_ = true;
-  }
-
-  // check age of input
-  if (is_input_ready_ && (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec() > 0.1) {
-    ROS_WARN_THROTTLE(1.0, "[%s]: input too old (%.4f), using zero input instead", getPrintName().c_str(),
-                      (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec());
-    is_input_ready_ = false;
-  }
-}
-/*//}*/
-
 /*//{ doCorrection() */
 void HdgGeneric::doCorrection(const Correction<hdg_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t &state_id) {
   doCorrection(meas.value, R, state_id, meas.stamp);
@@ -471,7 +435,7 @@ void HdgGeneric::doCorrection(const Correction<hdg_generic::n_measurements>::Mea
 /*//}*/
 
 /*//{ doCorrection() */
-void HdgGeneric::doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const ros::Time &meas_stamp) {
+void HdgGeneric::doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const rclcpp::Time &meas_stamp) {
 
   if (!isInitialized()) {
     return;
@@ -498,7 +462,7 @@ void HdgGeneric::doCorrection(const z_t &z, const double R, const StateId_t &H_i
     innovation_(0) = mrs_lib::geometry::radians::dist(mrs_lib::geometry::radians(meas(0)), mrs_lib::geometry::radians(sc.x(POSITION)));
 
     if (fabs(innovation_(0)) > pos_innovation_limit_) {
-      ROS_WARN_THROTTLE(1.0, "[%s]: innovation too large - hdg: %.2f", getPrintName().c_str(), innovation_(0));
+      RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: innovation too large - hdg: %.2f", getPrintName().c_str(), innovation_(0));
       innovation_ok_ = false;
       changeState(ERROR_STATE);
     }
@@ -521,7 +485,7 @@ void HdgGeneric::doCorrection(const z_t &z, const double R, const StateId_t &H_i
   }
   catch (const std::exception &e) {
     // In case of error, alert the user
-    ROS_ERROR("[%s]: LKF correction failed: %s", getPrintName().c_str(), e.what());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: LKF correction failed: %s", getPrintName().c_str(), e.what());
   }
 
   mrs_lib::set_mutexed(mutex_sc_, sc, sc_);
@@ -631,32 +595,32 @@ void HdgGeneric::setInputCoeff(const double &input_coeff) {
 /*//{ generateRepredictorModels() */
 void HdgGeneric::generateRepredictorModels(const double input_coeff) {
 
-    for (int i = 0; i < hdg_generic::n_states; i++) {
+  for (int i = 0; i < hdg_generic::n_states; i++) {
 
-      auto lambda_generateA = [input_coeff](const double dt) {
-        A_t A;
-        // clang-format off
+    auto lambda_generateA = [input_coeff](const double dt) {
+      A_t A;
+      // clang-format off
         A <<
           1, dt,
           0, 1-(input_coeff * dt);
-        // clang-format on
-        return A;
-      };
+      // clang-format on
+      return A;
+    };
 
-      auto lambda_generateB = [input_coeff]([[maybe_unused]] const double dt) {
-        B_t B = B.Zero();
-        // clang-format off
+    auto lambda_generateB = [input_coeff]([[maybe_unused]] const double dt) {
+      B_t B = B.Zero();
+      // clang-format off
         B <<
           0,
           (input_coeff * dt);
-        // clang-format on
-        return B;
-      };
+      // clang-format on
+      return B;
+    };
 
-      H_t H = H_t::Zero();
-      H(i)  = 1;
-      models_.push_back(std::make_shared<varstep_lkf_t>(lambda_generateA, lambda_generateB, H));
-    }
+    H_t H = H_t::Zero();
+    H(i)  = 1;
+    models_.push_back(std::make_shared<varstep_lkf_t>(lambda_generateA, lambda_generateB, H));
+  }
 }
 /*//}*/
 
@@ -682,23 +646,55 @@ void HdgGeneric::generateB() {
 }
 /*//}*/
 
+/* callbackParameters() //{ */
+
+rcl_interfaces::msg::SetParametersResult HdgGeneric::callbackParameters(std::vector<rclcpp::Parameter> parameters) {
+
+  rcl_interfaces::msg::SetParametersResult result;
+
+  // Note that setting a parameter to a nonsensical value (such as setting the `param_namespace.floating_number` parameter to `hello`)
+  // doesn't have any effect - it doesn't even call this callback.
+  for (auto &param : parameters) {
+
+    RCLCPP_INFO_STREAM(node_->get_logger(), "got parameter: '" << param.get_name() << "' with value '" << param.value_to_string() << "'");
+
+    if (param.get_name() == "pos") {
+
+      auto Q = mrs_lib::get_mutexed(mtx_Q_, Q_);
+
+      Q(POSITION, POSITION) = param.as_double();
+
+      mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+
+    } else if (param.get_name() == "vel") {
+
+      auto Q = mrs_lib::get_mutexed(mtx_Q_, Q_);
+
+      Q(VELOCITY, VELOCITY) = param.as_double();
+
+      mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+
+    } else {
+
+      RCLCPP_WARN_STREAM(node_->get_logger(), "parameter: '" << param.get_name() << "' is not dynamically reconfigurable!");
+      result.successful = false;
+      result.reason     = "Parameter '" + param.get_name() + "' is not dynamically reconfigurable!";
+      return result;
+    }
+  }
+
+  RCLCPP_INFO(node_->get_logger(), "params updated");
+  result.successful = true;
+  result.reason     = "OK";
+
+  return result;
+}
+
+//}
+
 /*//{ getLastValidHdg() */
 double HdgGeneric::getLastValidHdg() const {
   return mrs_lib::get_mutexed(mutex_last_valid_hdg_, last_valid_hdg_);
-}
-/*//}*/
-
-/*//{ callbackReconfigure() */
-void HdgGeneric::callbackReconfigure(HeadingEstimatorConfig &config, [[maybe_unused]] uint32_t level) {
-
-  if (!isInitialized()) {
-    return;
-  }
-
-  Q_t Q;
-  Q(POSITION, POSITION) = config.pos;
-  Q(VELOCITY, VELOCITY) = config.vel;
-  mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
 }
 /*//}*/
 
