@@ -10,7 +10,10 @@ namespace mrs_uav_state_estimators
 {
 
 /* initialize() //{*/
-void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) {
+void LatGeneric::initialize(const rclcpp::Node::SharedPtr &node, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) {
+
+  node_  = node;
+  clock_ = node->get_clock();
 
   ch_ = ch;
   ph_ = ph;
@@ -29,9 +32,8 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
     1, 0, 0, 0, 0, 0,
     0, 1, 0, 0, 0, 0;
 
-  innovation_ << 
+  innovation_ <<
     0, 0;
-
 
   // clang-format on
 
@@ -39,8 +41,10 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
   if (is_core_plugin_) {
 
-    ph->param_loader->addYamlFile(ros::package::getPath(package_name_) + "/config/private/" + parent_state_est_name_ + "/" + getName() + ".yaml");
-    ph->param_loader->addYamlFile(ros::package::getPath(package_name_) + "/config/public/" + parent_state_est_name_ + "/" + getName() + ".yaml");
+    ph->param_loader->addYamlFile(ament_index_cpp::get_package_share_directory(package_name_) + "/config/private/" + parent_state_est_name_ + "/" + getName() +
+                                  ".yaml");
+    ph->param_loader->addYamlFile(ament_index_cpp::get_package_share_directory(package_name_) + "/config/public/" + parent_state_est_name_ + "/" + getName() +
+                                  ".yaml");
   }
 
   ph->param_loader->setPrefix(ch_->package_name + "/" + Support::toSnakeCase(ch_->nodelet_name) + "/" + getNamespacedName() + "/");
@@ -61,7 +65,7 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
   for (auto corr_name : correction_names_) {
     corrections_.push_back(std::make_shared<Correction<lat_generic::n_measurements>>(
-        nh, getNamespacedName(), corr_name, ns_frame_id_, EstimatorType_t::LATERAL, ch_, ph_, [this](int a, int b) { return this->getState(a, b); },
+        node, getNamespacedName(), corr_name, ns_frame_id_, EstimatorType_t::LATERAL, ch_, ph_, [this](int a, int b) { return this->getState(a, b); },
         [this](const Correction<lat_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t state) {
           return this->doCorrection(meas, R, state);
         }));
@@ -69,8 +73,8 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
   // | ------- check if all parameters loaded successfully ------ |
   if (!ph->param_loader->loadedSuccessfully()) {
-    ROS_ERROR("[%s]: Could not load all non-optional parameters. Shutting down.", getPrintName().c_str());
-    ros::shutdown();
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: Could not load all non-optional parameters. Shutting down.", getPrintName().c_str());
+    rclcpp::shutdown();
   }
 
   ph->param_loader->setPrefix(ch_->package_name + "/" + Support::toSnakeCase(ch_->nodelet_name) + "/" + getNamespacedName() + "/");
@@ -89,14 +93,63 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
   Q_(stateIdToIndex(ACCELERATION, AXIS_Y), stateIdToIndex(ACCELERATION, AXIS_Y)) = tmp_noise;
 
   // | ------------- initialize dynamic reconfigure ------------- |
-  drmgr_ =
-      std::make_unique<drmgr_t>(ros::NodeHandle("~/" + getNamespacedName()), true, getPrintName(), boost::bind(&LatGeneric::callbackReconfigure, this, _1, _2));
-  drmgr_->config.pos = Q_(stateIdToIndex(POSITION, AXIS_X), stateIdToIndex(POSITION, AXIS_X));
-  drmgr_->config.vel = Q_(stateIdToIndex(VELOCITY, AXIS_X), stateIdToIndex(VELOCITY, AXIS_X));
-  drmgr_->config.acc = Q_(stateIdToIndex(ACCELERATION, AXIS_X), stateIdToIndex(ACCELERATION, AXIS_X));
-  drmgr_->update_config(drmgr_->config);
+
+  // original DRS from ROS1
+  /* drmgr_ = */
+  /*     std::make_unique<drmgr_t>(ros::NodeHandle("~/" + getNamespacedName()), true, getPrintName(), boost::bind(&LatGeneric::callbackReconfigure, this, _1,
+   * _2)); */
+  /* drmgr_->config.pos = Q_(stateIdToIndex(POSITION, AXIS_X), stateIdToIndex(POSITION, AXIS_X)); */
+  /* drmgr_->config.vel = Q_(stateIdToIndex(VELOCITY, AXIS_X), stateIdToIndex(VELOCITY, AXIS_X)); */
+  /* drmgr_->config.acc = Q_(stateIdToIndex(ACCELERATION, AXIS_X), stateIdToIndex(ACCELERATION, AXIS_X)); */
+  /* drmgr_->update_config(drmgr_->config); */
+
+  {
+    auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+
+    rcl_interfaces::msg::FloatingPointRange range;
+
+    range.from_value = 0.0;
+    range.to_value   = 100000.0;
+
+    param_desc.floating_point_range = {range};
+
+    param_desc.type = rcl_interfaces::msg::ParameterType::PARAMETER_DOUBLE;
+
+    node_->declare_parameter("pos", 0.0, param_desc);
+  }
+
+  {
+    auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+
+    rcl_interfaces::msg::FloatingPointRange range;
+
+    range.from_value = 0.0;
+    range.to_value   = 100000.0;
+
+    param_desc.floating_point_range = {range};
+
+    param_desc.type = rcl_interfaces::msg::ParameterType::PARAMETER_DOUBLE;
+
+    node_->declare_parameter("vel", 0.0, param_desc);
+  }
+
+  {
+    auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+
+    rcl_interfaces::msg::FloatingPointRange range;
+
+    range.from_value = 0.0;
+    range.to_value   = 100000.0;
+
+    param_desc.floating_point_range = {range};
+
+    param_desc.type = rcl_interfaces::msg::ParameterType::PARAMETER_DOUBLE;
+
+    node_->declare_parameter("acc", 0.0, param_desc);
+  }
 
   // | --------------- Kalman filter intialization -------------- |
+
   const x_t        x0 = x_t::Zero();
   const P_t        P0 = 1e3 * P_t::Identity();
   const statecov_t sc0({x0, P0});
@@ -107,49 +160,60 @@ void LatGeneric::initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHan
 
     generateRepredictorModels(input_coeff_);
 
-    const u_t       u0 = u_t::Zero();
-    const ros::Time t0 = ros::Time::now();
-    lkf_rep_           = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
+    const u_t          u0 = u_t::Zero();
+    const rclcpp::Time t0 = clock_->now();
+    lkf_rep_              = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_.at(0), rep_buffer_size_);
 
     setDt(1.0 / ch_->desired_uav_state_rate);
   }
 
-  // | ------------------ timers initialization ----------------- |
-  timer_update_ = nh.createTimer(ros::Rate(ch_->desired_uav_state_rate), &LatGeneric::timerUpdate, this); 
-  /* timer_check_health_       = nh.createTimer(ros::Rate(ch_->desired_uav_state_rate), &LatGeneric::timerCheckHealth, this); */
+  // | ------------------------- timers ------------------------- |
+
+  mrs_lib::TimerHandlerOptions opts;
+
+  opts.node      = node_;
+  opts.autostart = true;
+
+  {
+    std::function<void()> callback_fcn = std::bind(&LatGeneric::timerUpdate, this);
+
+    timer_update_ = std::make_shared<mrs_lib::ROSTimer>(opts, rclcpp::Rate(ch_->desired_uav_state_rate, clock_), callback_fcn);
+
+    timer_update_last_time_ = rclcpp::Time(0, 0, clock_->get_clock_type());
+  }
 
   // | --------------- subscribers initialization --------------- |
   // subscriber to odometry
-  mrs_lib::SubscribeHandlerOptions shopts;
-  shopts.nh                 = nh;
+  mrs_lib::SubscriberHandlerOptions shopts;
+
+  shopts.node               = node;
   shopts.node_name          = getPrintName();
   shopts.no_message_timeout = mrs_lib::no_timeout;
   shopts.threadsafe         = true;
   shopts.autostart          = true;
-  shopts.queue_size         = 10;
-  shopts.transport_hints    = ros::TransportHints().tcpNoDelay();
 
-  sh_control_input_ = mrs_lib::SubscribeHandler<mrs_msgs::EstimatorInput>(shopts, "control_input_in");
-  sh_hdg_state_ =
-      mrs_lib::SubscribeHandler<mrs_msgs::EstimatorOutput>(shopts, hdg_source_topic_);  // for transformation of desired accelerations from body to global frame
+  sh_control_input_ = mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorInput>(shopts, "control_input_in");
+
+  // for transformation of desired accelerations from body to global frame
+  sh_hdg_state_ = mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorOutput>(shopts, hdg_source_topic_);
 
   // | ---------------- publishers initialization --------------- |
   if (ch_->debug_topics.input) {
-    ph_input_ = mrs_lib::PublisherHandler<mrs_msgs::Float64ArrayStamped>(nh, getNamespacedName() + "/input", 10);
+    ph_input_ = mrs_lib::PublisherHandler<mrs_msgs::msg::Float64ArrayStamped>(node, "~/" + getNamespacedName() + "/input");
   }
   if (ch_->debug_topics.output) {
-    ph_output_ = mrs_lib::PublisherHandler<mrs_msgs::EstimatorOutput>(nh, getNamespacedName() + "/output", 10);
+    ph_output_ = mrs_lib::PublisherHandler<mrs_msgs::msg::EstimatorOutput>(node, "~/" + getNamespacedName() + "/output");
   }
   if (ch_->debug_topics.diag) {
-    ph_diagnostics_ = mrs_lib::PublisherHandler<mrs_msgs::EstimatorDiagnostics>(nh, getNamespacedName() + "/diagnostics", 10);
+    ph_diagnostics_ = mrs_lib::PublisherHandler<mrs_msgs::msg::EstimatorDiagnostics>(node, "~/" + getNamespacedName() + "/diagnostics");
   }
 
   // | ------------------ finish initialization ----------------- |
 
   if (changeState(INITIALIZED_STATE)) {
-    ROS_INFO("[%s]: Estimator initialized, version %s", getPrintName().c_str(), VERSION);
+    RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator initialized, version %s", getPrintName().c_str(), VERSION);
   } else {
-    ROS_INFO("[%s]: Estimator could not be initialized", getPrintName().c_str());
+    RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator could not be initialized", getPrintName().c_str());
   }
 }
 /*//}*/
@@ -163,7 +227,7 @@ bool LatGeneric::start(void) {
     return true;
 
   } else {
-    ROS_WARN_THROTTLE(1.0, "[%s]: Estimator must be in READY_STATE to start it", getPrintName().c_str());
+    RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator must be in READY_STATE to start it", getPrintName().c_str());
     return false;
   }
 }
@@ -186,7 +250,7 @@ bool LatGeneric::pause(void) {
 bool LatGeneric::reset(void) {
 
   if (!isInitialized()) {
-    ROS_ERROR("[%s]: Cannot reset uninitialized estimator", getPrintName().c_str());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: Cannot reset uninitialized estimator", getPrintName().c_str());
     return false;
   }
 
@@ -207,20 +271,20 @@ bool LatGeneric::reset(void) {
   lkf_ = std::make_shared<lkf_t>(A_, B_, H_);
   if (is_repredictor_enabled_) {
 
-    const u_t       u0 = u_t::Zero();
-    const ros::Time t0 = ros::Time(0);
-    lkf_rep_           = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_[0], rep_buffer_size_);
+    const u_t          u0 = u_t::Zero();
+    const rclcpp::Time t0 = rclcpp::Time(0);
+    lkf_rep_              = std::make_unique<mrs_lib::Repredictor<varstep_lkf_t>>(x0, P0, u0, Q_, t0, models_[0], rep_buffer_size_);
   }
 
   changeState(INITIALIZED_STATE);
-  ROS_INFO("[%s]: Estimator reset", getPrintName().c_str());
+  RCLCPP_INFO(node_->get_logger(), "[%s]: Estimator reset", getPrintName().c_str());
 
   return true;
 }
 /*//}*/
 
 /* timerUpdate() //{*/
-void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
+void LatGeneric::timerUpdate() {
 
   if (!isInitialized()) {
     return;
@@ -229,39 +293,44 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
   switch (getCurrentSmState()) {
 
     case UNINITIALIZED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
       break;
     }
 
     case READY_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
       break;
     }
 
     case INITIALIZED_STATE: {
+
       // initialize the estimator with current corrections
       for (auto correction : corrections_) {
+
         auto res = correction->getProcessedCorrection();
+
         if (res) {
-          auto measurement_stamped = res.value(); 
+          auto measurement_stamped = res.value();
           setState(measurement_stamped.value(AXIS_X), correction->getStateId(), AXIS_X);
           setState(measurement_stamped.value(AXIS_Y), correction->getStateId(), AXIS_Y);
-          ROS_INFO_THROTTLE(1.0, "[%s]: Setting initial state to: %.2f %.2f", getPrintName().c_str(), measurement_stamped.value(AXIS_X),
-                            measurement_stamped.value(AXIS_Y));
+          RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Setting initial state to: %.2f %.2f", getPrintName().c_str(),
+                               measurement_stamped.value(AXIS_X), measurement_stamped.value(AXIS_Y));
         } else {
-          ROS_INFO_THROTTLE(1.0, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(), correction->getNamespacedName().c_str());
+          RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(),
+                               correction->getNamespacedName().c_str());
           return;
         }
       }
-      ROS_INFO_THROTTLE(1.0, "[%s]: Ready to start", getPrintName().c_str());
+
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Ready to start", getPrintName().c_str());
       changeState(READY_STATE);
       break;
     }
 
     case STARTED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
       if (isConverged()) {
-        ROS_INFO_THROTTLE(1.0, "[%s]: LKF converged", getPrintName().c_str());
+        RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: LKF converged", getPrintName().c_str());
         changeState(RUNNING_STATE);
       }
       break;
@@ -270,7 +339,8 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
     case RUNNING_STATE: {
       for (auto correction : corrections_) {
         if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
+          RCLCPP_ERROR_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Correction %s is not healthy!", getPrintName().c_str(),
+                                correction->getNamespacedName().c_str());
           changeState(ERROR_STATE);
         }
       }
@@ -278,35 +348,36 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
     }
 
     case STOPPED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is stopped", getPrintName().c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator is stopped", getPrintName().c_str());
       break;
     }
 
     case ERROR_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
+      RCLCPP_INFO_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
 
-      ros::Time t_now = ros::Time::now();
+      rclcpp::Time t_now = clock_->now();
       if (is_error_state_first_time_) {
         prev_time_in_error_state_  = t_now;
         is_error_state_first_time_ = false;
-        error_state_duration_      = ros::Duration(0.0);
+        error_state_duration_      = 0;
       }
-      error_state_duration_ += t_now - prev_time_in_error_state_;
+      error_state_duration_ += t_now.seconds() - prev_time_in_error_state_.seconds();
 
 
       // check if all corrections are healthy now
       bool all_corrections_healthy = true;
       for (auto correction : corrections_) {
         if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
+          RCLCPP_ERROR_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: Correction %s is not healthy!", getPrintName().c_str(),
+                                correction->getNamespacedName().c_str());
           all_corrections_healthy = false;
         }
       }
 
       if (all_corrections_healthy && innovation_ok_) {
         // initialize the estimator again if corrections become healthy
-        if (error_state_duration_.toSec() > 5.0) {
-          ROS_INFO("[%s]: corrections healthy for %.2f s", getPrintName().c_str(), error_state_duration_.toSec());
+        if (error_state_duration_ > 5.0) {
+          RCLCPP_INFO(node_->get_logger(), "[%s]: corrections healthy for %.2f s", getPrintName().c_str(), error_state_duration_);
           changeState(INITIALIZED_STATE);
           is_error_state_first_time_ = true;
         }
@@ -325,9 +396,9 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
   }
 
   // check age of input
-  if (is_input_ready_ && (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec() > 0.1) {  // TODO: parametrize, if older than say 1 second, eland
-    ROS_WARN_THROTTLE(1.0, "[%s]: input too old (%.4f s), using zero input instead", getPrintName().c_str(),
-                      (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec());
+  if (is_input_ready_ && (clock_->now() - sh_control_input_.lastMsgTime()).seconds() > 0.1) {  // TODO: parametrize, if older than say 1 second, eland
+    RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: input too old (%.4f s), using zero input instead", getPrintName().c_str(),
+                         (clock_->now() - sh_control_input_.lastMsgTime()).seconds());
     is_input_ready_ = false;
   }
 
@@ -345,37 +416,42 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
   }
 
   // obtain dt for state prediction
-  double dt = (event.current_real - event.last_real).toSec();
-  if (dt <= 0.0 || dt > 1.0) {  // sometimes the timer ticks twice simultaneously in simulation - we ignore the second tick, in case of stopping and starting the timer, the last_real is 0
+  double dt = (clock_->now() - timer_update_last_time_).seconds();
+
+  timer_update_last_time_ = clock_->now();
+
+  if (dt <= 0.0 || dt > 1.0) {  // sometimes the timer ticks twice simultaneously in simulation - we ignore the second tick, in case of stopping and starting
+                                // the timer, the last_real is 0
     return;
   }
 
-  if (!is_repredictor_enabled_) { // repredictor calculates dt on its own
+  if (!is_repredictor_enabled_) {  // repredictor calculates dt on its own
     setDt(dt);
   }
 
   // obtain unbiased desired control acceleration in the estimator frame that will be used as input to the estimator
-  u_t       u;
-  ros::Time input_stamp;
+  u_t          u;
+  rclcpp::Time input_stamp;
   if (is_input_ready_ && is_hdg_state_ready_) {
     auto res = fun_get_hdg_();
     if (!res) {
-      ROS_ERROR_THROTTLE(1.0, "[%s]: could not obtain heading", getPrintName().c_str());
+      RCLCPP_ERROR_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: could not obtain heading", getPrintName().c_str());
       return;
     }
 
     const tf2::Vector3 des_acc_global = getAccGlobal(sh_control_input_.getMsg(), res.value());
     input_stamp                       = sh_control_input_.getMsg()->header.stamp;
-    if (input_coeff_ != default_input_coeff_){
+    if (input_coeff_ != default_input_coeff_) {
       setInputCoeff(default_input_coeff_);
     }
     u(0) = des_acc_global.getX();
     u(1) = des_acc_global.getY();
 
   } else {  // this is ok before the controller starts controlling but bad during actual flight (causes delayed estimated acceleration and velocity)
-    ROS_DEBUG_THROTTLE(1.0, "[%s]: not receiving control input, estimation suboptimal, potentially unstable", getPrintName().c_str());
-    input_stamp = ros::Time::now();
-    if (input_coeff_ != 0){
+    RCLCPP_DEBUG_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: not receiving control input, estimation suboptimal, potentially unstable",
+                          getPrintName().c_str());
+    input_stamp = clock_->now();
+    if (input_coeff_ != 0) {
       setInputCoeff(0);
     }
     u = u_t::Zero();
@@ -400,13 +476,13 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
     std::scoped_lock lock(mutex_lkf_);
     if (is_repredictor_enabled_) {
       lkf_rep_->addInputChangeWithNoise(u, Q, input_stamp, models_[0]);
-      sc = lkf_rep_->predictTo(ros::Time::now());
+      sc = lkf_rep_->predictTo(clock_->now());
     } else {
       sc = lkf_->predict(sc, u, Q, dt_);
     }
   }
   catch (const std::exception &e) {
-    ROS_ERROR("[%s]: LKF prediction failed: %s", getPrintName().c_str(), e.what());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: LKF prediction failed: %s", getPrintName().c_str(), e.what());
     changeState(ERROR_STATE);
   }
 
@@ -420,104 +496,6 @@ void LatGeneric::timerUpdate(const ros::TimerEvent &event) {
 }
 /*//}*/
 
-/*//{ timerCheckHealth() */
-void LatGeneric::timerCheckHealth([[maybe_unused]] const ros::TimerEvent &event) {
-
-  if (!isInitialized()) {
-    return;
-  }
-
-  switch (getCurrentSmState()) {
-
-    case UNINITIALIZED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s initialization", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      break;
-    }
-
-    case READY_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s estimator start", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      break;
-    }
-
-    case INITIALIZED_STATE: {
-      // initialize the estimator with current corrections
-      for (auto correction : corrections_) {
-        auto res = correction->getProcessedCorrection();
-        if (res) {
-          auto measurement_stamped = res.value();
-          setState(measurement_stamped.value(AXIS_X), correction->getStateId(), AXIS_X);
-          setState(measurement_stamped.value(AXIS_Y), correction->getStateId(), AXIS_Y);
-          ROS_INFO_THROTTLE(1.0, "[%s]: Setting initial state to: %.2f %.2f", getPrintName().c_str(), measurement_stamped.value(AXIS_X),
-                            measurement_stamped.value(AXIS_Y));
-        } else {
-          ROS_INFO_THROTTLE(1.0, "[%s]: %s correction %s", getPrintName().c_str(), Support::waiting_for_string.c_str(), correction->getPrintName().c_str());
-          return;
-        }
-      }
-      ROS_INFO_THROTTLE(1.0, "[%s]: Ready to start", getPrintName().c_str());
-      changeState(READY_STATE);
-      break;
-    }
-
-    case STARTED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: %s convergence of LKF", getPrintName().c_str(), Support::waiting_for_string.c_str());
-      if (isConverged()) {
-        ROS_INFO_THROTTLE(1.0, "[%s]: LKF converged", getPrintName().c_str());
-        changeState(RUNNING_STATE);
-      }
-      break;
-    }
-
-    case RUNNING_STATE: {
-      for (auto correction : corrections_) {
-        if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
-          changeState(ERROR_STATE);
-        }
-      }
-      break;
-    }
-
-    case STOPPED_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is stopped", getPrintName().c_str());
-      break;
-    }
-
-    case ERROR_STATE: {
-      ROS_INFO_THROTTLE(1.0, "[%s]: Estimator is in ERROR state", getPrintName().c_str());
-      bool all_corrections_healthy = true;
-      for (auto correction : corrections_) {
-        if (!correction->isHealthy()) {
-          ROS_ERROR_THROTTLE(1.0, "[%s]: Correction %s is not healthy!", getPrintName().c_str(), correction->getNamespacedName().c_str());
-          correction->resetProcessors();
-          all_corrections_healthy = false;
-        }
-      }
-      // initialize the estimator again if corrections become healthy
-      if (all_corrections_healthy) {
-        changeState(INITIALIZED_STATE);
-      }
-      break;
-    }
-  }
-
-  if (sh_control_input_.newMsg()) {
-    is_input_ready_ = true;
-  }
-
-  // check age of input
-  if (is_input_ready_ && (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec() > 0.1) {  // TODO: parametrize, if older than say 1 second, eland
-    ROS_WARN_THROTTLE(1.0, "[%s]: input too old (%.4f s), using zero input instead", getPrintName().c_str(),
-                      (ros::Time::now() - sh_control_input_.lastMsgTime()).toSec());
-    is_input_ready_ = false;
-  }
-
-  if (fun_get_hdg_()) {
-    is_hdg_state_ready_ = true;
-  }
-}
-/*//}*/
-
 /*//{ doCorrection() */
 void LatGeneric::doCorrection(const Correction<lat_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t &state_id) {
   doCorrection(meas.value, R, state_id, meas.stamp);
@@ -525,7 +503,7 @@ void LatGeneric::doCorrection(const Correction<lat_generic::n_measurements>::Mea
 /*//}*/
 
 /*//{ doCorrection() */
-void LatGeneric::doCorrection(const z_t &z, const double R, const StateId_t &state_id, const ros::Time &meas_stamp) {
+void LatGeneric::doCorrection(const z_t &z, const double R, const StateId_t &state_id, const rclcpp::Time &meas_stamp) {
 
   if (!isInitialized()) {
     return;
@@ -546,24 +524,24 @@ void LatGeneric::doCorrection(const z_t &z, const double R, const StateId_t &sta
       innovation_(1)      = z(1) - getState(POSITION, AXIS_Y);
 
       if (fabs(innovation_(0)) > pos_innovation_limit_ || fabs(innovation_(1)) > pos_innovation_limit_) {
-        ROS_WARN_THROTTLE(1.0, "[%s]: innovation too large - [%.2f %.2f] lim: %.2f", getPrintName().c_str(), innovation_(0), innovation_(1),
-                          pos_innovation_limit_);
+        RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: innovation too large - [%.2f %.2f] lim: %.2f", getPrintName().c_str(), innovation_(0),
+                             innovation_(1), pos_innovation_limit_);
         innovation_ok_ = false;
         switch (exc_innovation_action_) {
           case ExcInnoAction_t::ELAND: {
-            ROS_WARN_THROTTLE(1.0, "[%s]: large innovation should trigger eland in control manager", getPrintName().c_str());
+            RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: large innovation should trigger eland in control manager", getPrintName().c_str());
             changeState(ERROR_STATE);
             break;
           }
           case ExcInnoAction_t::SWITCH: {
-            ROS_WARN_THROTTLE(1.0, "[%s]: innovation should trigger estimator switch but no eland", getPrintName().c_str());
+            RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: innovation should trigger estimator switch but no eland", getPrintName().c_str());
             innovation_(0) = 0.0;  // this is quite hacky but is there other way to switch estimators and not trigger eland by the large innovation?
             innovation_(1) = 0.0;
             changeState(ERROR_STATE);
             break;
           }
           case ExcInnoAction_t::MITIGATE: {
-            ROS_WARN_THROTTLE(1.0, "[%s]: large innovation should trigger estimate jump mitigation", getPrintName().c_str());
+            RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: large innovation should trigger estimate jump mitigation", getPrintName().c_str());
             innovation_(0)      = 0.0;  // this is quite hacky but is there other way to switch estimators and not trigger eland by the large innovation?
             innovation_(1)      = 0.0;
             is_mitigating_jump_ = true;
@@ -572,7 +550,7 @@ void LatGeneric::doCorrection(const z_t &z, const double R, const StateId_t &sta
             break;
           }
           case ExcInnoAction_t::NONE: {
-            ROS_WARN_THROTTLE(1.0, "[%s]: large innovation ignored", getPrintName().c_str());
+            RCLCPP_WARN_THROTTLE(node_->get_logger(), *clock_, 1000, "[%s]: large innovation ignored", getPrintName().c_str());
             break;
           }
         }
@@ -599,7 +577,7 @@ void LatGeneric::doCorrection(const z_t &z, const double R, const StateId_t &sta
   }
   catch (const std::exception &e) {
     // In case of error, alert the user
-    ROS_ERROR("[%s]: LKF correction failed: %s", getPrintName().c_str(), e.what());
+    RCLCPP_ERROR(node_->get_logger(), "[%s]: LKF correction failed: %s", getPrintName().c_str(), e.what());
   }
 
   mrs_lib::set_mutexed(mutex_sc_, sc, sc_);
@@ -708,11 +686,11 @@ void LatGeneric::setInputCoeff(const double &input_coeff) {
 
 /*//{ generateRepredictorModels() */
 void LatGeneric::generateRepredictorModels(const double input_coeff) {
-    for (int i = 0; (int)(i < lat_generic::n_states / 2); i++) {
+  for (int i = 0; (int)(i < lat_generic::n_states / 2); i++) {
 
-      auto lambda_generateA = [input_coeff](const double dt) {
-        A_t A;
-        // clang-format off
+    auto lambda_generateA = [input_coeff](const double dt) {
+      A_t A;
+      // clang-format off
         A <<
           1, 0, dt, 0, 0.5 * dt * dt, 0,
           0, 1, 0, dt, 0, 0.5 * dt * dt,
@@ -720,13 +698,13 @@ void LatGeneric::generateRepredictorModels(const double input_coeff) {
           0, 0, 0, 1, 0, dt,
           0, 0, 0, 0, 1-(input_coeff * dt), 0,
           0, 0, 0, 0, 0, 1-(input_coeff * dt);
-        // clang-format on
-        return A;
-      };
+      // clang-format on
+      return A;
+    };
 
-      auto lambda_generateB = [input_coeff]([[maybe_unused]] const double dt) {
-        B_t B = B.Zero();
-        // clang-format off
+    auto lambda_generateB = [input_coeff]([[maybe_unused]] const double dt) {
+      B_t B = B.Zero();
+      // clang-format off
           B <<
             0, 0,
             0, 0,
@@ -734,16 +712,16 @@ void LatGeneric::generateRepredictorModels(const double input_coeff) {
             0, 0,
             input_coeff * dt, 0,
             0, input_coeff * dt;
-        // clang-format on
+      // clang-format on
 
-        return B;
-      };
+      return B;
+    };
 
-      H_t H                = H.Zero();
-      H(AXIS_X, i * 2)     = 1;
-      H(AXIS_Y, i * 2 + 1) = 1;
-      models_.push_back(std::make_shared<varstep_lkf_t>(lambda_generateA, lambda_generateB, H));
-    }
+    H_t H                = H.Zero();
+    H(AXIS_X, i * 2)     = 1;
+    H(AXIS_Y, i * 2 + 1) = 1;
+    models_.push_back(std::make_shared<varstep_lkf_t>(lambda_generateA, lambda_generateB, H));
+  }
 }
 /*//}*/
 
@@ -777,23 +755,62 @@ void LatGeneric::generateB() {
 }
 /*//}*/
 
-/*//{ callbackReconfigure() */
-void LatGeneric::callbackReconfigure(LateralEstimatorConfig &config, [[maybe_unused]] uint32_t level) {
+/* callbackParameters() //{ */
 
-  if (!isInitialized()) {
-    return;
+rcl_interfaces::msg::SetParametersResult LatGeneric::callbackParameters(std::vector<rclcpp::Parameter> parameters) {
+
+  rcl_interfaces::msg::SetParametersResult result;
+
+  // Note that setting a parameter to a nonsensical value (such as setting the `param_namespace.floating_number` parameter to `hello`)
+  // doesn't have any effect - it doesn't even call this callback.
+  for (auto &param : parameters) {
+
+    RCLCPP_INFO_STREAM(node_->get_logger(), "got parameter: '" << param.get_name() << "' with value '" << param.value_to_string() << "'");
+
+    if (param.get_name() == "pos") {
+
+      auto Q = mrs_lib::get_mutexed(mtx_Q_, Q_);
+
+      Q(stateIdToIndex(POSITION, AXIS_X), stateIdToIndex(POSITION, AXIS_X)) = param.as_double();
+      Q(stateIdToIndex(POSITION, AXIS_Y), stateIdToIndex(POSITION, AXIS_Y)) = param.as_double();
+
+      mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+
+    } else if (param.get_name() == "vel") {
+
+      auto Q = mrs_lib::get_mutexed(mtx_Q_, Q_);
+
+      Q(stateIdToIndex(VELOCITY, AXIS_X), stateIdToIndex(VELOCITY, AXIS_X)) = param.as_double();
+      Q(stateIdToIndex(VELOCITY, AXIS_Y), stateIdToIndex(VELOCITY, AXIS_Y)) = param.as_double();
+
+      mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+
+    } else if (param.get_name() == "acc") {
+
+      auto Q = mrs_lib::get_mutexed(mtx_Q_, Q_);
+
+      Q(stateIdToIndex(ACCELERATION, AXIS_X), stateIdToIndex(ACCELERATION, AXIS_X)) = param.as_double();
+      Q(stateIdToIndex(ACCELERATION, AXIS_Y), stateIdToIndex(ACCELERATION, AXIS_Y)) = param.as_double();
+
+      mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+
+    } else {
+
+      RCLCPP_WARN_STREAM(node_->get_logger(), "parameter: '" << param.get_name() << "' is not dynamically reconfigurable!");
+      result.successful = false;
+      result.reason     = "Parameter '" + param.get_name() + "' is not dynamically reconfigurable!";
+      return result;
+    }
   }
 
-  Q_t Q;
-  Q(stateIdToIndex(POSITION, AXIS_X), stateIdToIndex(POSITION, AXIS_X))         = config.pos;
-  Q(stateIdToIndex(POSITION, AXIS_Y), stateIdToIndex(POSITION, AXIS_Y))         = config.pos;
-  Q(stateIdToIndex(VELOCITY, AXIS_X), stateIdToIndex(VELOCITY, AXIS_X))         = config.vel;
-  Q(stateIdToIndex(VELOCITY, AXIS_Y), stateIdToIndex(VELOCITY, AXIS_Y))         = config.vel;
-  Q(stateIdToIndex(ACCELERATION, AXIS_X), stateIdToIndex(ACCELERATION, AXIS_X)) = config.acc;
-  Q(stateIdToIndex(ACCELERATION, AXIS_Y), stateIdToIndex(ACCELERATION, AXIS_Y)) = config.acc;
-  mrs_lib::set_mutexed(mtx_Q_, Q, Q_);
+  RCLCPP_INFO(node_->get_logger(), "params updated");
+  result.successful = true;
+  result.reason     = "OK";
+
+  return result;
 }
-/*//}*/
+
+//}
 
 /*//{ getNamespacedName() */
 std::string LatGeneric::getNamespacedName() const {
