@@ -3,23 +3,21 @@
 
 /* includes //{ */
 
-#include <ros/ros.h>
+#include <rclcpp/rclcpp.hpp>
 
-#include <nav_msgs/Odometry.h>
-#include <sensor_msgs/Range.h>
+#include <nav_msgs/msg/odometry.hpp>
+#include <sensor_msgs/msg/range.hpp>
 
 #include <mrs_lib/lkf.h>
 #include <mrs_lib/repredictor.h>
 #include <mrs_lib/profiler.h>
 #include <mrs_lib/param_loader.h>
-#include <mrs_lib/subscribe_handler.h>
-
-#include <functional>
+#include <mrs_lib/subscriber_handler.h>
 
 #include <mrs_uav_state_estimators/estimators/altitude/altitude_estimator.h>
 #include <mrs_uav_state_estimators/estimators/correction.h>
 
-#include <mrs_uav_state_estimators/AltitudeEstimatorConfig.h>
+#include <ament_index_cpp/get_package_share_directory.hpp>
 
 //}
 
@@ -39,8 +37,6 @@ class AltGeneric : public AltitudeEstimator<alt_generic::n_states> {
 
   const std::string package_name_ = "mrs_uav_state_estimators";
 
-  typedef mrs_lib::DynamicReconfigureMgr<AltitudeEstimatorConfig> drmgr_t;
-
   using lkf_t         = mrs_lib::LKF<alt_generic::n_states, alt_generic::n_inputs, alt_generic::n_measurements>;
   using varstep_lkf_t = mrs_lib::varstepLKF<alt_generic::n_states, alt_generic::n_inputs, alt_generic::n_measurements>;
   using A_t           = lkf_t::A_t;
@@ -58,6 +54,8 @@ class AltGeneric : public AltitudeEstimator<alt_generic::n_states> {
 
   using StateId_t = mrs_uav_managers::estimation_manager::StateId_t;
 
+  rcl_interfaces::msg::SetParametersResult callbackParameters(std::vector<rclcpp::Parameter> parameters);
+
 private:
   std::string parent_state_est_name_;
 
@@ -74,15 +72,12 @@ private:
   statecov_t                                  sc_;
   mutable std::mutex                          mutex_sc_;
 
-  std::unique_ptr<drmgr_t> drmgr_;
-  void                     callbackReconfigure(AltitudeEstimatorConfig &config, [[maybe_unused]] uint32_t level);
-
   z_t                innovation_;
   mutable std::mutex mtx_innovation_;
 
-  bool          is_error_state_first_time_ = true;
-  ros::Duration error_state_duration_;
-  ros::Time     prev_time_in_error_state_;
+  bool         is_error_state_first_time_ = true;
+  double       error_state_duration_;
+  rclcpp::Time prev_time_in_error_state_;
 
   bool is_repredictor_enabled_;
   int  rep_buffer_size_ = 200;
@@ -92,18 +87,19 @@ private:
   std::vector<std::string>                                              correction_names_;
   std::vector<std::shared_ptr<Correction<alt_generic::n_measurements>>> corrections_;
 
-  mrs_lib::SubscribeHandler<mrs_msgs::EstimatorInput> sh_control_input_;
-  void                                                timeoutCallback(const std::string &topic, const ros::Time &last_msg);
-  std::atomic<bool>                                   is_input_ready_ = false;
+  mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorInput> sh_control_input_;
+  void                                                      timeoutCallback(const std::string &topic, const rclcpp::Time &last_msg);
+  std::atomic<bool>                                         is_input_ready_ = false;
 
-  ros::Timer timer_update_;
-  void       timerUpdate(const ros::TimerEvent &event);
+  std::shared_ptr<mrs_lib::ROSTimer> timer_update_;
+  void                               timerUpdate();
+  rclcpp::Time                       timer_update_last_time_;
 
-  ros::Timer timer_check_health_;
-  void       timerCheckHealth(const ros::TimerEvent &event);
+  std::shared_ptr<mrs_lib::ROSTimer> timer_check_health_;
+  void                               timerCheckHealth();
 
   void doCorrection(const Correction<alt_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t &state_id);
-  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const ros::Time &meas_stamp);
+  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const rclcpp::Time &meas_stamp);
 
   bool isConverged();
 
@@ -111,14 +107,18 @@ private:
   mutable std::mutex mtx_Q_;
 
 public:
-  AltGeneric(const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name, const bool is_core_plugin)
-      : AltitudeEstimator<alt_generic::n_states>(name, ns_frame_id), parent_state_est_name_(parent_state_est_name), is_core_plugin_(is_core_plugin) {
+  AltGeneric(const rclcpp::Node::SharedPtr &node, const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name,
+             const bool is_core_plugin)
+      : AltitudeEstimator<alt_generic::n_states>(node, name, ns_frame_id), parent_state_est_name_(parent_state_est_name), is_core_plugin_(is_core_plugin) {
+
+    node_  = node;
+    clock_ = node->get_clock();
   }
 
   ~AltGeneric(void) {
   }
 
-  void initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
+  void initialize(const rclcpp::Node::SharedPtr &node, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
   bool start(void) override;
   bool pause(void) override;
   bool reset(void) override;
@@ -149,8 +149,8 @@ public:
   void generateA();
   void generateB();
 
-  void timeoutOdom(const std::string &topic, const ros::Time &last_msg, const int n_pubs);
-  void timeoutRange(const std::string &topic, const ros::Time &last_msg, const int n_pubs);
+  void timeoutOdom(const std::string &topic, const rclcpp::Time &last_msg, const int n_pubs);
+  void timeoutRange(const std::string &topic, const rclcpp::Time &last_msg, const int n_pubs);
 
   std::string getNamespacedName() const;
 
