@@ -3,21 +3,19 @@
 
 /* includes //{ */
 
-#include <ros/ros.h>
+#include <rclcpp/rclcpp.hpp>
 
-#include <nav_msgs/Odometry.h>
+#include <nav_msgs/msg/odometry.hpp>
 
 #include <mrs_lib/lkf.h>
 #include <mrs_lib/repredictor.h>
 #include <mrs_lib/profiler.h>
 #include <mrs_lib/param_loader.h>
-#include <mrs_lib/subscribe_handler.h>
+#include <mrs_lib/subscriber_handler.h>
 #include <mrs_lib/geometry/cyclic.h>
 
 #include <mrs_uav_state_estimators/estimators/heading/heading_estimator.h>
 #include <mrs_uav_state_estimators/estimators/correction.h>
-
-#include <mrs_uav_state_estimators/HeadingEstimatorConfig.h>
 
 //}
 
@@ -36,8 +34,6 @@ const int n_measurements = 1;
 class HdgGeneric : public HeadingEstimator<hdg_generic::n_states> {
 
   const std::string package_name_ = "mrs_uav_state_estimators";
-
-  typedef mrs_lib::DynamicReconfigureMgr<HeadingEstimatorConfig> drmgr_t;
 
   using lkf_t         = mrs_lib::LKF<hdg_generic::n_states, hdg_generic::n_inputs, hdg_generic::n_measurements>;
   using varstep_lkf_t = mrs_lib::varstepLKF<hdg_generic::n_states, hdg_generic::n_inputs, hdg_generic::n_measurements>;
@@ -73,9 +69,6 @@ private:
   statecov_t                                  sc_;
   mutable std::mutex                          mutex_sc_;
 
-  std::unique_ptr<drmgr_t> drmgr_;
-  void                     callbackReconfigure(HeadingEstimatorConfig &config, [[maybe_unused]] uint32_t level);
-
   z_t                innovation_;
   mutable std::mutex mtx_innovation_;
 
@@ -87,18 +80,16 @@ private:
   std::vector<std::string>                                              correction_names_;
   std::vector<std::shared_ptr<Correction<hdg_generic::n_measurements>>> corrections_;
 
-  mrs_lib::SubscribeHandler<mrs_msgs::EstimatorInput> sh_control_input_;
-  void                                                timeoutCallback(const std::string &topic, const ros::Time &last_msg);
-  std::atomic<bool>                                   is_input_ready_ = false;
+  mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorInput> sh_control_input_;
+  void                                                      timeoutCallback(const std::string &topic, const rclcpp::Time &last_msg);
+  std::atomic<bool>                                         is_input_ready_ = false;
 
-  ros::Timer timer_update_;
-  void       timerUpdate(const ros::TimerEvent &event);
-
-  ros::Timer timer_check_health_;
-  void       timerCheckHealth(const ros::TimerEvent &event);
+  std::shared_ptr<mrs_lib::ROSTimer> timer_update_;
+  void                               timerUpdate();
+  rclcpp::Time                       timer_update_last_time_;
 
   void doCorrection(const Correction<hdg_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t &state_id);
-  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const ros::Time &meas_stamp);
+  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const rclcpp::Time &meas_stamp);
 
   bool isConverged();
 
@@ -108,15 +99,18 @@ private:
   mutable std::mutex mutex_last_valid_hdg_;
   double             last_valid_hdg_;
 
+  rcl_interfaces::msg::SetParametersResult callbackParameters(std::vector<rclcpp::Parameter> parameters);
+
 public:
-  HdgGeneric(const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name, const bool is_core_plugin)
-      : HeadingEstimator<hdg_generic::n_states>(name, ns_frame_id), parent_state_est_name_(parent_state_est_name), is_core_plugin_(is_core_plugin) {
+  HdgGeneric(const rclcpp::Node::SharedPtr &node, const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name,
+             const bool is_core_plugin)
+      : HeadingEstimator<hdg_generic::n_states>(node, name, ns_frame_id), parent_state_est_name_(parent_state_est_name), is_core_plugin_(is_core_plugin) {
   }
 
   ~HdgGeneric(void) {
   }
 
-  void initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
+  void initialize(const rclcpp::Node::SharedPtr &node, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
   bool start(void) override;
   bool pause(void) override;
   bool reset(void) override;
@@ -148,7 +142,6 @@ public:
 
   void generateA();
   void generateB();
-
 
   std::string getNamespacedName() const;
 
