@@ -19,7 +19,8 @@ template <int n_states>
 class HeadingEstimator : public PartialEstimator<n_states, 1> {
 
 protected:
-  HeadingEstimator(const std::string& name, const std::string& frame_id) : PartialEstimator<n_states, 1>(heading::type, name, frame_id) {
+  HeadingEstimator(const rclcpp::Node::SharedPtr& node, const std::string& name, const std::string& frame_id)
+      : PartialEstimator<n_states, 1>(node, heading::type, name, frame_id) {
   }
 
   mutable std::mutex mutex_last_valid_hdg_;
