@@ -3,22 +3,20 @@
 
 /* includes //{ */
 
-#include <ros/ros.h>
+#include <rclcpp/rclcpp.hpp>
 
-#include <nav_msgs/Odometry.h>
+#include <nav_msgs/msg/odometry.hpp>
 
-#include <sensor_msgs/Imu.h>
+#include <sensor_msgs/msg/imu.hpp>
 
 #include <mrs_lib/lkf.h>
 #include <mrs_lib/repredictor.h>
 #include <mrs_lib/profiler.h>
 #include <mrs_lib/param_loader.h>
-#include <mrs_lib/subscribe_handler.h>
+#include <mrs_lib/subscriber_handler.h>
 
 #include <mrs_uav_state_estimators/estimators/lateral/lateral_estimator.h>
 #include <mrs_uav_state_estimators/estimators/correction.h>
-
-#include <mrs_uav_state_estimators/LateralEstimatorConfig.h>
 
 //}
 
@@ -35,8 +33,6 @@ const int n_measurements = 2;
 }  // namespace lat_generic
 
 class LatGeneric : public LateralEstimator<lat_generic::n_states> {
-
-  typedef mrs_lib::DynamicReconfigureMgr<LateralEstimatorConfig> drmgr_t;
 
   using lkf_t         = mrs_lib::LKF<lat_generic::n_states, lat_generic::n_inputs, lat_generic::n_measurements>;
   using varstep_lkf_t = mrs_lib::varstepLKF<lat_generic::n_states, lat_generic::n_inputs, lat_generic::n_measurements>;
@@ -71,15 +67,12 @@ private:
   statecov_t                                  sc_;
   mutable std::mutex                          mutex_sc_;
 
-  std::unique_ptr<drmgr_t> drmgr_;
-  void                     callbackReconfigure(LateralEstimatorConfig &config, [[maybe_unused]] uint32_t level);
-
   z_t                innovation_;
   mutable std::mutex mtx_innovation_;
 
-  bool          is_error_state_first_time_ = true;
-  ros::Duration error_state_duration_;
-  ros::Time     prev_time_in_error_state_;
+  bool         is_error_state_first_time_ = true;
+  double       error_state_duration_;
+  rclcpp::Time prev_time_in_error_state_;
 
   bool is_repredictor_enabled_;
   int  rep_buffer_size_ = 200;
@@ -89,35 +82,34 @@ private:
   std::vector<std::string>                                              correction_names_;
   std::vector<std::shared_ptr<Correction<lat_generic::n_measurements>>> corrections_;
 
-  mrs_lib::SubscribeHandler<mrs_msgs::EstimatorInput> sh_control_input_;
-  void                                                timeoutCallback(const std::string &topic, const ros::Time &last_msg);
-  std::atomic<bool>                                   is_input_ready_ = false;
+  mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorInput> sh_control_input_;
+  void                                                      timeoutCallback(const std::string &topic, const rclcpp::Time &last_msg);
+  std::atomic<bool>                                         is_input_ready_ = false;
 
 
-  std::function<std::optional<double>()>               fun_get_hdg_;
-  std::string                                          hdg_source_topic_;
-  mrs_lib::SubscribeHandler<mrs_msgs::EstimatorOutput> sh_hdg_state_;
-  std::atomic<bool>                                    is_hdg_state_ready_ = false;
+  std::function<std::optional<double>()>                     fun_get_hdg_;
+  std::string                                                hdg_source_topic_;
+  mrs_lib::SubscriberHandler<mrs_msgs::msg::EstimatorOutput> sh_hdg_state_;
+  std::atomic<bool>                                          is_hdg_state_ready_ = false;
 
-  ros::Timer timer_update_;
-  void       timerUpdate(const ros::TimerEvent &event);
-
-  ros::Timer timer_check_health_;
-  void       timerCheckHealth(const ros::TimerEvent &event);
-
+  std::shared_ptr<mrs_lib::ROSTimer> timer_update_;
+  void                               timerUpdate();
+  rclcpp::Time                       timer_update_last_time_;
 
   void doCorrection(const Correction<lat_generic::n_measurements>::MeasurementStamped &meas, const double R, const StateId_t &state_id);
-  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const ros::Time &meas_stamp);
+  void doCorrection(const z_t &z, const double R, const StateId_t &H_idx, const rclcpp::Time &meas_stamp);
 
   bool isConverged();
 
   Q_t                getQ();
   mutable std::mutex mtx_Q_;
 
+  rcl_interfaces::msg::SetParametersResult callbackParameters(std::vector<rclcpp::Parameter> parameters);
+
 public:
-  LatGeneric(const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name, const bool is_core_plugin,
-             std::function<std::optional<double>()> fun_get_hdg)
-      : LateralEstimator<lat_generic::n_states>(name, ns_frame_id),
+  LatGeneric(const rclcpp::Node::SharedPtr &node, const std::string &name, const std::string &ns_frame_id, const std::string &parent_state_est_name,
+             const bool is_core_plugin, std::function<std::optional<double>()> fun_get_hdg)
+      : LateralEstimator<lat_generic::n_states>(node, name, ns_frame_id),
         parent_state_est_name_(parent_state_est_name),
         is_core_plugin_(is_core_plugin),
         fun_get_hdg_(fun_get_hdg) {
@@ -126,7 +118,7 @@ public:
   ~LatGeneric(void) {
   }
 
-  void initialize(ros::NodeHandle &nh, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
+  void initialize(const rclcpp::Node::SharedPtr &node, const std::shared_ptr<CommonHandlers_t> &ch, const std::shared_ptr<PrivateHandlers_t> &ph) override;
   bool start(void) override;
   bool pause(void) override;
   bool reset(void) override;
