@@ -4,13 +4,13 @@
 
 #include <mrs_uav_state_estimators/processors/processor.h>
 
+#include <mrs_uav_managers/estimation_manager/support.h>
+
 #include <mrs_lib/param_loader.h>
-#include <mrs_lib/subscribe_handler.h>
+#include <mrs_lib/subscriber_handler.h>
 #include <mrs_lib/attitude_converter.h>
 
 #include <Eigen/Dense>
-
-#include <limits>
 
 namespace mrs_uav_state_estimators
 {
@@ -24,8 +24,8 @@ public:
   typedef Eigen::Matrix<double, n_measurements, 1> measurement_t;
 
 public:
-  ProcExcessiveTilt(ros::NodeHandle& nh, const std::string& correction_name, const std::string& name, const std::shared_ptr<CommonHandlers_t>& ch,
-                    const std::shared_ptr<PrivateHandlers_t>& ph);
+  ProcExcessiveTilt(const rclcpp::Node::SharedPtr& node, const std::string& correction_name, const std::string& name,
+                    const std::shared_ptr<CommonHandlers_t>& ch, const std::shared_ptr<PrivateHandlers_t>& ph);
 
   std::tuple<bool, bool> process(measurement_t& measurement) override;
   void                   reset();
@@ -33,15 +33,15 @@ public:
 private:
   double max_tilt_sq_;
 
-  std::string                                                 orientation_topic_;
-  mrs_lib::SubscribeHandler<geometry_msgs::QuaternionStamped> sh_orientation_;
+  std::string                                                       orientation_topic_;
+  mrs_lib::SubscriberHandler<geometry_msgs::msg::QuaternionStamped> sh_orientation_;
 };
 
 /*//{ constructor */
 template <int n_measurements>
-ProcExcessiveTilt<n_measurements>::ProcExcessiveTilt(ros::NodeHandle& nh, const std::string& correction_name, const std::string& name,
+ProcExcessiveTilt<n_measurements>::ProcExcessiveTilt(const rclcpp::Node::SharedPtr& node, const std::string& correction_name, const std::string& name,
                                                      const std::shared_ptr<CommonHandlers_t>& ch, const std::shared_ptr<PrivateHandlers_t>& ph)
-    : Processor<n_measurements>(nh, correction_name, name, ch, ph) {
+    : Processor<n_measurements>(node, correction_name, name, ch, ph) {
 
   // | --------------------- load parameters -------------------- |
   ph->param_loader->setPrefix(ch->package_name + "/" + Support::toSnakeCase(ch->nodelet_name) + "/" + Processor<n_measurements>::getNamespacedName() + "/");
@@ -55,21 +55,21 @@ ProcExcessiveTilt<n_measurements>::ProcExcessiveTilt(ros::NodeHandle& nh, const 
   max_tilt_sq_ = std::pow(max_tilt, 2);
 
   if (!ph->param_loader->loadedSuccessfully()) {
-    ROS_ERROR("[%s]: Could not load all non-optional parameters. Shutting down.", Processor<n_measurements>::getPrintName().c_str());
-    ros::shutdown();
+    RCLCPP_ERROR(this->node_->get_logger(), "[%s]: Could not load all non-optional parameters. Shutting down.",
+                 Processor<n_measurements>::getPrintName().c_str());
+    rclcpp::shutdown();
   }
 
   // | -------------- initialize subscribe handlers ------------- |
-  mrs_lib::SubscribeHandlerOptions shopts;
-  shopts.nh                 = nh;
+  mrs_lib::SubscriberHandlerOptions shopts;
+
+  shopts.node               = node;
   shopts.node_name          = Processor<n_measurements>::getPrintName();
-  shopts.no_message_timeout = ros::Duration(1.0);
+  shopts.no_message_timeout = rclcpp::Duration(std::chrono::duration<double>(1.0));
   shopts.threadsafe         = true;
   shopts.autostart          = true;
-  shopts.queue_size         = 10;
-  shopts.transport_hints    = ros::TransportHints().tcpNoDelay();
 
-  sh_orientation_ = mrs_lib::SubscribeHandler<geometry_msgs::QuaternionStamped>(shopts, "/" + ch->uav_name + "/" + orientation_topic_);
+  sh_orientation_ = mrs_lib::SubscriberHandler<geometry_msgs::msg::QuaternionStamped>(shopts, "/" + ch->uav_name + "/" + orientation_topic_);
 }
 /*//}*/
 
@@ -96,13 +96,15 @@ std::tuple<bool, bool> ProcExcessiveTilt<n_measurements>::process([[maybe_unused
     const bool is_excessive_tilt = std::pow(tilt, 2) > max_tilt_sq_;
 
     if (is_excessive_tilt) {
-      ROS_WARN_THROTTLE(1.0, "[%s]: excessive tilt of %.2f deg. Not fusing correction.", Processor<n_measurements>::getPrintName().c_str(), tilt / M_PI * 180);
+      RCLCPP_WARN_THROTTLE(this->node_->get_logger(), *this->clock_, 1000, "[%s]: excessive tilt of %.2f deg. Not fusing correction.",
+                           Processor<n_measurements>::getPrintName().c_str(), tilt / M_PI * 180);
       ok_flag     = false;
       should_fuse = false;
     }
   }
   catch (...) {
-    ROS_ERROR_THROTTLE(1.0, "[%s]: failed obtaining tilt value", Processor<n_measurements>::getPrintName().c_str());
+    RCLCPP_ERROR_THROTTLE(this->node_->get_logger(), *this->clock_, 1000, "[%s]: failed obtaining tilt value",
+                          Processor<n_measurements>::getPrintName().c_str());
     ok_flag     = false;
     should_fuse = false;
   }
